+++ src/pages/ApplicationsOverview.tsx
@@ -159,6 +159,7 @@ export class ApplicationsOverview extends BasePage {
             events: this.generateEventsLink(),
             logs: this.generateLogsLink(),
             traces: this.generateTracesLink(),
+            services: this.generateServicesLink(),
             edit: this.generateEditLink(),
             config: `/plugins/${APP_ID}`,
           }}