+++ src/pages/ClusterStatus.tsx
@@ -72,6 +72,7 @@ export class ClusterStatus extends BasePage {
             events: this.generateEventsLink(),
             logs: this.generateLogsLink(),
             traces: this.generateTracesLink(),
+            services: this.generateServicesLink(),
             edit: this.generateEditLink(),
             config: `/plugins/${APP_ID}`,
           }}