+++ src/pages/Events.tsx
@@ -86,6 +86,7 @@ export class EventsPage extends BasePage {
             events: this.generateEventsLink(),
             logs: this.generateLogsLink(),
             traces: this.generateTracesLink(),
+            services: this.generateServicesLink(),
             edit: this.generateEditLink(),
             config: `/plugins/${APP_ID}`,
           }}